+++ out/compilers/compilerBase.js
@@ -154,10 +154,59 @@ class CompilerBase {
             return true;
         });
     }
+    VerifyCompilerFilesExistsAsync() {
+        return __awaiter(this, void 0, void 0, function* () {
+            console.log('debugger:CompilerBase.VerifyCompilerFilesExistsAsync');
+            // Prepare
+            let compilerFileList = this.GetCompilerFileList();
+            let result = true;
+            // Process
+            for (let compilerFileName of compilerFileList) {
+                // Prepare
+                let compilerFilePath = path.join(this.FolderOrPath, compilerFileName);
+                // Validate
+                if (!(yield filesystem.FileExistsAsync(compilerFilePath))) {
+                    // Not found
+                    application.WriteToCompilerTerminal(`ERROR: Unable to locate compiler file '${compilerFileName}'. `);
+                    result = false;
+                }
+            }
+            // Failed?, 
+            if (!result) {
+                let message = "NOTE: your anti-virus software may have quarantined one or more files related to the compiler due to a false/positive test and where this is the case please ensure you whitelist to allow these files to used.  Alternatively try re-installing the extension.";
+                application.WriteToCompilerTerminal(message);
+            }
+            // Result
+            return result;
+        });
+    }
     RepairFilePermissionsAsync() {
         return __awaiter(this, void 0, void 0, function* () {
-            return true;
+            console.log('debugger:CompilerBase.RepairFilePermissionsAsync');
+            // Validate
+            if (this.CustomFolderOrPath || application.IsWindows) {
+                return true;
+            }
+            // Prepare
+            let compilerFileList = this.GetCompilerFileList();
+            let result = true;
+            // Process
+            for (let compilerFileName of compilerFileList) {
+                // Prepare
+                let compilerFilePath = path.join(this.FolderOrPath, compilerFileName);
+                // Validate
+                if (!(yield filesystem.ChModAsync(compilerFilePath))) {
+                    // Not found
+                    application.WriteToCompilerTerminal(`WARNING: Unable to set file permissions for compiler file '${compilerFileName}'. `);
+                    result = false;
+                }
+            }
+            // Result
+            return result;
         });
+    }
+    GetCompilerFileList() {
+        return [];
     }
     LoadConfigurationAsync() {
         return __awaiter(this, void 0, void 0, function* () {